decouple TokensInliner
